Move getStatus impl into FlowNode, fix NPE.
