+++ src/main/java/org/jenkinsci/plugins/workflow/graph/AtomNode.java
@@ -24,14 +24,8 @@
 
 package org.jenkinsci.plugins.workflow.graph;
 
-import hudson.model.Result;
-import org.jenkinsci.plugins.workflow.actions.ErrorAction;
-import org.jenkinsci.plugins.workflow.actions.FlowNodeStatusAction;
 import org.jenkinsci.plugins.workflow.flow.FlowExecution;
-import org.jenkinsci.plugins.workflow.steps.FlowInterruptedException;
 import org.jenkinsci.plugins.workflow.steps.Step;
-
-import javax.annotation.CheckForNull;
 
 /**
  * FlowNode that has no further FlowNodes inside.
@@ -42,28 +36,5 @@ import javax.annotation.CheckForNull;
 public abstract class AtomNode extends FlowNode {
     protected AtomNode(FlowExecution exec, String id, FlowNode... parents) {
         super(exec, id, parents);
-    }
-
-    @CheckForNull
-    @Override
-    public Result getStatus() {
-        if (isActive()) {
-            return null;
-        } else {
-            ErrorAction errorAction = getError();
-            if (errorAction != null) {
-                if (errorAction.getError() instanceof FlowInterruptedException) {
-                    return Result.ABORTED;
-                } else {
-                    return Result.FAILURE;
-                }
-            }
-            FlowNodeStatusAction statusAction = getPersistentAction(FlowNodeStatusAction.class);
-            if (statusAction != null) {
-                return statusAction.getResult();
-            }
-
-            return Result.SUCCESS;
-        }
     }
 }
